requirements.txt
PySide6==6.5.0
qt-material==2.14
onnxruntime-gpu
huggingface-hub==0.15.1
pillow==9.5.0
numpy
pandas
opencv-contrib-python
